+++ src/test/scala/is/hail/expr/ir/MatrixIRSuite.scala
@@ -1,0 +1,45 @@
+package is.hail.expr.ir
+
+import is.hail.SparkSuite
+import is.hail.expr.ir.TestUtils._
+import is.hail.variant.MatrixTable
+import org.apache.spark.sql.Row
+import org.testng.annotations.Test
+
+class MatrixIRSuite extends SparkSuite {
+
+  def rangeMatrix: MatrixIR = MatrixTable.range(hc, 20, 20, Some(4)).ast
+
+  @Test def testScanCountBehavesLikeIndex() {
+    val mt = rangeMatrix
+    val oldRow = Ref("va", mt.typ.rvRowType)
+
+    val newRow = InsertFields(oldRow, Seq("idx" -> IRScanCount))
+
+    val newMatrix = MatrixMapRows(mt, newRow, None)
+    val rows = new MatrixTable(hc, newMatrix).rowsTable().collect()
+    assert(rows.forall { case Row(row_idx, idx) => row_idx == idx})
+  }
+
+  @Test def testScanCollectBehavesLikeRange() {
+    val mt = rangeMatrix
+    val oldRow = Ref("va", mt.typ.rvRowType)
+
+    val newRow = InsertFields(oldRow, Seq("range" -> IRScanCollect(GetField(oldRow, "row_idx"))))
+
+    val newMatrix = MatrixMapRows(mt, newRow, None)
+    val rows = new MatrixTable(hc, newMatrix).rowsTable().collect()
+    assert(rows.forall { case Row(row_idx: Int, range: IndexedSeq[Int]) => range sameElements Array.range(0, row_idx)})
+  }
+
+  @Test def testScanCollectBehavesLikeRangeWithAggregation() {
+    val mt = rangeMatrix
+    val oldRow = Ref("va", mt.typ.rvRowType)
+
+    val newRow = InsertFields(oldRow, Seq("range" -> IRScanCollect(IRAggCount * GetField(oldRow, "row_idx").toL)))
+
+    val newMatrix = MatrixMapRows(mt, newRow, None)
+    val rows = new MatrixTable(hc, newMatrix).rowsTable().collect()
+    assert(rows.forall { case Row(row_idx: Int, range: IndexedSeq[Int]) => range sameElements Array.range(0, row_idx).map(_ * 20)})
+  }
+}
